rename BaseTokenGenerateEngine to SQLTokenGeneratorBuilder
